Refactor value types (#93)
- Changes the template logic to give a dictionary type enum value instead of inner details
- The need for a persistence choice is hidden in a lower layer.
- Value Stores are separated by 3 different modes of writing: compile, merge, append merge.
- Implement support for append merge of string dictionaries.

fixes #93

+++ keyvi/include/keyvi/vector/vector_file.h
@@ -66,7 +66,8 @@ class VectorFile {
 
     in_stream.seekg(KEYVI_VECTOR_BEGIN_LEN);
     const auto file_ptree = util::SerializationUtils::ReadJsonRecord(in_stream);
-    if (value_store_type != boost::lexical_cast<size_t>(file_ptree.get<std::string>(VALUE_STORE_TYPE_LABEl))) {
+    if (static_cast<int>(value_store_type) !=
+        boost::lexical_cast<int>(file_ptree.get<std::string>(VALUE_STORE_TYPE_LABEl))) {
       throw std::invalid_argument("wrong vector file");
     }
     manifest_ = file_ptree.get<std::string>(MANIFEST_LABEL);
@@ -96,7 +97,7 @@ class VectorFile {
     boost::property_tree::ptree file_ptree;
     file_ptree.put("file_version", "1");
     file_ptree.put(MANIFEST_LABEL, manifest);
-    file_ptree.put(VALUE_STORE_TYPE_LABEl, value_store->GetValueStoreType());
+    file_ptree.put(VALUE_STORE_TYPE_LABEl, static_cast<int>(value_store->GetValueStoreType()));
     file_ptree.put("index_version", "1");
 
     boost::property_tree::ptree index_ptree;
